+++ src/main/java/org/threadly/concurrent/AbstractSubmitterExecutor.java
@@ -9,9 +9,10 @@ import org.threadly.util.ArgumentVerifier;
 /**
  * <p>Since the conversion to a {@link SubmitterExecutor} from an executor is often the same (just 
  * using the {@link ListenableFutureTask} to wrap the task).  This class provides an easy way to 
- * create a {@link SubmitterExecutor}.  Take a look at {@link ExecutorWrapper} for an easy example 
- * of how this is used.  In general this wont be useful outside of Threadly developers, but must 
- * be a public interface since it is used in sub-packages.</p>
+ * create a {@link SubmitterExecutor}.  Take a look at 
+ * {@link org.threadly.concurrent.wrapper.ExecutorWrapper} for an easy example of how this is 
+ * used.  In general this wont be useful outside of Threadly developers, but must be a public 
+ * visibility since it is used in sub-packages.</p>
  * 
  * <p>If you do find yourself using this class, please post an issue on github to tell us why.  If 
  * there is something you want our schedulers to provide, we are happy to hear about it.</p>
