+++ src/main/java/org/threadly/concurrent/wrapper/ExecutorWrapper.java
@@ -1,28 +1,37 @@
-package org.threadly.concurrent;
+package org.threadly.concurrent.wrapper;
 
 import java.util.concurrent.Executor;
 
+import org.threadly.concurrent.AbstractSubmitterExecutor;
+import org.threadly.util.ArgumentVerifier;
+
 /**
  * <p>A simple wrapper class for {@link Executor} implementations to provide 
- * {@link SubmitterExecutor} capabilities.</p>
+ * {@link org.threadly.concurrent.SubmitterExecutor} capabilities.</p>
  * 
  * <p>In addition this implementation returns 
  * {@link org.threadly.concurrent.future.ListenableFuture} future implementations, making it an 
  * easy way to convert your favorite executor to use ListenableFutures.</p>
  * 
- * @deprecated Moved to {@link org.threadly.concurrent.wrapper.ExecutorWrapper}
- * 
  * @author jent - Mike Jensen
  * @since 1.0.0
  */
-@Deprecated
-public class ExecutorWrapper extends org.threadly.concurrent.wrapper.ExecutorWrapper {
+public class ExecutorWrapper extends AbstractSubmitterExecutor {
+  protected final Executor executor;
+  
   /**
    * Constructors a new wrapper instance with the provided executor to defer calls to.
    * 
    * @param executor {@link Executor} instance.
    */
   public ExecutorWrapper(Executor executor) {
-    super(executor);
+    ArgumentVerifier.assertNotNull(executor, "executor");
+    
+    this.executor = executor;
+  }
+
+  @Override
+  protected void doExecute(Runnable task) {
+    executor.execute(task);
   }
 }
